activityOneWeek13/requirements.txt
Flask>=2.2
openai>=1.0.0
python-dotenv>=0.21.0
requests>=2.28
pytest>=7.0
